pyproject.toml
[tool.poetry]
name = "PyTorchYolo"
version = "1.8.0"
readme = "README.md"
repository = "https://github.com/eriklindernoren/PyTorch-YOLOv3"
description = "Minimal PyTorch implementation of YOLO"
authors = [
    "Florian Vahl <git@flova.de>",
    "Erik Linder-Noren <eriklindernoren@gmail.com>",
]
license = "GPL-3.0"

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
torch = ">=1.10.1, < 1.13.0"
torchvision = ">=0.13.1"
matplotlib = "^3.3.3"
tensorboard = "^2.10.0"
terminaltables = "^3.1.0"
Pillow = "^9.1.0"
tqdm = "^4.64.1"
urllib3 = [
    { version = "<=1.22", python = ">=3.8,<3.9" },
    { version = "^1.23", python = ">=3.9" },
] # Temp pin because of crash issue
scipy = [
    { version = "<=1.6", python = ">=3.8,<3.9" },
    { version = "^1.9", python = ">=3.9,<4.0" },
]
imgaug = "^0.4.0"
torchsummary = "^1.5.1"
numpy = "^1.23.4"

[tool.poetry.dev-dependencies]
profilehooks = "^1.12.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
yolo-detect = "pytorchyolo.detect:run"
yolo-train = "pytorchyolo.train:run"
yolo-test = "pytorchyolo.test:run"
